Cairo update
Update code to Cairo v0.10.0 and starknet.js v4.8.0.
Use local Devnet.
Improvement of README.
