+++ .env
@@ -1,0 +1,9 @@
+# with starknet-devnet --seed 0 , available account 0 is :
+OZ_ACCOUNT_ADDRESS=0x7e00d496e324876bbc8531f2d9a82bf154d1a04a50218ee74cdd372f75a551a
+export OZ_ACCOUNT_ADDRESS
+OZ_ACCOUNT_PRIVATE_KEY=0xe3e70682c2094cac629f6fbed82c07cd
+export OZ_ACCOUNT_PRIVATE_KEY
+
+STARKNET_PROVIDER_BASE_URL=http://127.0.0.1:5050
+# https://alpha4.starknet.io for Goerli alpha testnet, https://alpha-mainnet.starknet.io for mainnet, http://127.0.0.1:5050 for local starknet-devnet
+export STARKNET_PROVIDER_BASE_URL
