+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "starknetjs-101",
+  "name": "starknet-js-typescript",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -10,7 +10,17 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "starknet": "^4.1.0"
+    "@toruslabs/starkware-crypto": "^1.1.0",
+    "axios": "^0.24.0",
+    "fs": "^0.0.1-security",
+    "starknet": "^4.8.0"
   },
-  "type": "module"
+  "devDependencies": {
+    "@types/elliptic": "^6.4.14",
+    "@types/node": "^16.11.10",
+    "@types/url-join": "^4.0.1",
+    "dotenv": "^16.0.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4"
+  }
 }
